feat(app-layout): add footer
